pyproject.toml
[project]
name = "reader3"
version = "0.1.0"
description = "Simple EPUB reader web app"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "ebooklib>=0.20",
    "fastapi>=0.121.2",
    "jinja2>=3.1.6",
    "uvicorn>=0.38.0",
    "deep-translator>=1.11.4",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "Pillow>=10.0.0",
]
